Ensure all pathname work. Rename repo to project for consistency.
