+++ app/[slug]/project/[projectId]/page.tsx
@@ -1,0 +1,7 @@
+export default async function Page({params}: {params: {projectId: string}}) {
+	return (
+		<div className='flex flex-col gap-4'>
+			<p>Coming soon...</p>
+		</div>
+	)
+}
