+++ app/[slug]/project/[projectId]/settings/page.tsx
